Revert "Fix more Win static analysis errors"
This reverts commit c56c0d9106d0e28658cf24b92c8cf8620e73d42b.

+++ src/libs/tagging.c
@@ -2526,11 +2526,11 @@ static void _import_button_clicked(GtkButton *button, dt_lib_module_t *self)
     char *filename = gtk_file_chooser_get_filename(GTK_FILE_CHOOSER(filechooser));
     char *dirname = g_path_get_dirname(filename);
     dt_conf_set_string("plugins/lighttable/tagging/last_import_export_location", dirname);
-    int count = dt_tag_import(filename);
+    ssize_t count = dt_tag_import(filename);
     if(count < 0)
       dt_control_log(_("error importing tags"));
     else
-      dt_control_log(_("%i tags imported"), count);
+      dt_control_log(_("%zd tags imported"), count);
     dt_free(filename);
     dt_free(dirname);
   }
